Add gasPrice calculation in runtime tx detail view
