+++ .changelog/1469.trivial.md
@@ -1,0 +1,1 @@
+Add gasPrice calculation in runtime tx detail view
