+++ src/app/components/Tokens/TokenTransfers.tsx
@@ -15,12 +15,13 @@ import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
 import { TokenTransferIcon } from './TokenTransferIcon'
 import { RoundedBalance } from '../RoundedBalance'
 import { useScreenSize } from '../../hooks/useScreensize'
-import { fromBaseUnits, getEthAccountAddressFromBase64 } from '../../utils/helpers'
+import { getEthAccountAddressFromBase64 } from '../../utils/helpers'
 import { TokenLink } from './TokenLink'
 import { PlaceholderLabel } from '../../utils/PlaceholderLabel'
 import { TokenTypeTag } from './TokenList'
 import { parseEvmEvent } from '../../utils/parseEvmEvent'
 import { Age } from '../Age'
+import { fromBaseUnits } from '../../utils/number-utils'
 
 const iconSize = '28px'
 const StyledCircle = styled(Box)(({ theme }) => ({
